pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
psycopg = {extras = ["binary"], version = "^3.1.19"}
geopandas = "^0.14.4"
pandas = "^2.2.2"
shapely = "^2.0.4"
pyogrio = "^0.8.0"
pyproj = "^3.6.1"
h3 = "^3.7.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
